Support binary responses using isBase64Encoded flag
API Gateway will treat lambda-proxy responses with `isBase64Encoded` set
to `true` as binary as long as the mime-type matches one of the allowed
binary types.

This change checks the flag and responds with a Buffer created from the
encoded string, instead of the string itself.

If one was being thorough one would probably also support this guy:
https://github.com/maciejtreder/serverless-apigw-binary

But I suspect the binary type configuration in APIG is an mostly an
additional layer of security?