fix: error feedback when using --useSeparateProcesses flag
When using the --useSeparateProcesses flag, the lambda function handler is loaded in a separate process, and serverless-offline communicates with this child process over IPC. When an unhandledException error occurs, the Error object wasn't being properly serialized. Then, due to the asynchronous loading of the handler code (require() in the ipcHelper), fatal errors were not being reported correctly to the request-client. This commit fixes these issues.