Proof of concept of assuming roles
Here's a proof of concept of assuming a role in serverless-offline. It basically works. If you're open to including in the library I'm happy to flesh out this PR to:
- make the behaviour optional
- read the role ARN from the function's serverless config
- adjust the code to live where you see fit; not knowing the codebase I've just put it in the first place I found that'd work

I and my team would really appreciate this feature. Lambda's running in AWS do not require profiles, so currently serverless-offline doesn't prefectly replicate lambda in this respect. Also, we are trying to encapsulate all our config in per-project files living within the project's repo. As AWS profiles can only be set in a global config file, finding a way to avoid using them woudl be very helpful