Reading headers defined on the proxy
Since it is possible to define extra headers on the Serverless proxy configuration level, this function will format those headers into an object. One example, is a header used to call a third party API, but we want it to be defined in the Serverless API only, hence not exposed in the Front End request.