fix: support handlers exported from nested modules
AWS currently allows handler functions to be buried deep inside of a
module or object but serverless-offline doesn't.

The change updates serverless offline to align with how AWS Lambda
attempts to find a handler function.

It also changes the fixture file
`src/lambda/__tests__/fixtures/lambdaFunction.fixture.js` to align with
how AWS expects the module path of handler to be.