feat: support Go on provided.al2 runtime
This adds support for the current runtime used for Go on AWS. The `go1.x` runtime is EOL as its underlying AMI became EOL on 2020-12-31 and `provided.al2` is its replacement (https://aws.amazon.com/blogs/compute/migrating-aws-lambda-functions-to-al2/). Whilst the AMI is still under extended maintenance support, this extends only until 2023-12-31 (https://aws.amazon.com/blogs/aws/update-on-amazon-linux-ami-end-of-life/).

This also ensures that any `vendor/` directories from Go in the runtime tests are ignored.